:recycle: [repositories] Move `provide` and `UnknownLocationError` to `registry`
