+++ src/cutty/repositories/domain/providers.py
@@ -7,7 +7,6 @@ from typing import Optional
 
 from yarl import URL
 
-from cutty.errors import CuttyError
 from cutty.filesystems.adapters.disk import DiskFilesystem
 from cutty.filesystems.domain.filesystem import Filesystem
 from cutty.filesystems.domain.path import Path
@@ -39,24 +38,6 @@ class Provider:
         self, location: Location, revision: Optional[Revision]
     ) -> Optional[Repository]:
         """Return the repository at the given location."""
-
-
-@dataclass
-class UnknownLocationError(CuttyError):
-    """The repository location could not be processed by any provider."""
-
-    location: Location
-
-
-def provide(
-    providers: Iterable[Provider], location: Location, revision: Optional[Revision]
-) -> Repository:
-    """Provide the repository located at the given URL."""
-    for provider in providers:
-        if repository := provider(location, revision):
-            return repository
-
-    raise UnknownLocationError(location)
 
 
 GetRevision = Callable[[pathlib.Path, Optional[Revision]], Optional[Revision]]
