+++ src/cutty/repositories/domain/registry.py
@@ -1,23 +1,43 @@
 """The provider registry is the main entry point of cutty.repositories."""
+from collections.abc import Iterable
 from collections.abc import Iterator
 from collections.abc import Mapping
+from dataclasses import dataclass
 from typing import Optional
 
 from yarl import URL
 
+from cutty.errors import CuttyError
 from cutty.filesystems.domain.path import Path
 from cutty.filesystems.domain.pathfs import PathFilesystem
 from cutty.filesystems.domain.purepath import PurePath
 from cutty.repositories.domain.fetchers import FetchMode
 from cutty.repositories.domain.locations import Location
 from cutty.repositories.domain.locations import parselocation
-from cutty.repositories.domain.providers import provide
 from cutty.repositories.domain.providers import Provider
 from cutty.repositories.domain.providers import ProviderFactory
 from cutty.repositories.domain.providers import ProviderName
 from cutty.repositories.domain.providers import ProviderStore
 from cutty.repositories.domain.providers import Repository
 from cutty.repositories.domain.revisions import Revision
+
+
+@dataclass
+class UnknownLocationError(CuttyError):
+    """The repository location could not be processed by any provider."""
+
+    location: Location
+
+
+def provide(
+    providers: Iterable[Provider], location: Location, revision: Optional[Revision]
+) -> Repository:
+    """Provide the repository located at the given URL."""
+    for provider in providers:
+        if repository := provider(location, revision):
+            return repository
+
+    raise UnknownLocationError(location)
 
 
 class ProviderRegistry:
