+++ tests/unit/repositories/domain/test_providers.py
@@ -19,11 +19,11 @@ from cutty.repositories.domain.locations import Location
 from cutty.repositories.domain.mounters import unversioned_mounter
 from cutty.repositories.domain.providers import constproviderfactory
 from cutty.repositories.domain.providers import LocalProvider
-from cutty.repositories.domain.providers import provide
 from cutty.repositories.domain.providers import Provider
 from cutty.repositories.domain.providers import ProviderStore
 from cutty.repositories.domain.providers import remoteproviderfactory
 from cutty.repositories.domain.providers import Repository
+from cutty.repositories.domain.registry import provide
 from cutty.repositories.domain.registry import ProviderRegistry
 from cutty.repositories.domain.revisions import Revision
 from cutty.repositories.domain.stores import Store
